Kubernetes/Docker-RappiOesteCGI/RappiOesteFlask/requirements.txt
Click==7.0
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
uWSGI==2.0.18
Werkzeug==0.15.3
mysql-connector